engine/pyproject.toml
[project]
name = "synthdata-engine"
version = "0.1.0"
description = "SynthData engine — synthetic NLP dataset generation pipeline"
readme = "README.md"
authors = [
    { name = "BeastxD7", email = "shashankdevadiga2003@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.94",
    "datasketch>=1.10.0",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "typer>=0.24.2",
]

[project.scripts]
synthdata = "synthdata_engine.cli:main"

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
